Widened ServerInfo.Build to 64-bits and added the BuildNumberConverter class
* 12-byte string values are assumed to be hexadecimal Git commit
numbers; for Splunk [6.3,)
* All other string values are assumed to be unsigned 32-bit decimal
integer values; Splunk [,6.3)

+++ src/Splunk.Client/Splunk/Client/IServerInfo.cs
@@ -14,10 +14,6 @@
  * under the License.
  */
 
-//// TODO:
-//// [O] Contracts
-//// [O] Documentation
-
 namespace Splunk.Client
 {
     using System;
@@ -54,7 +50,7 @@ namespace Splunk.Client
         /// <value>
         /// The build.
         /// </value>
-        int Build
+        long Build
         { get; }
 
         /// <summary>
